+++ lib/flag.dart
@@ -5,28 +5,22 @@ import 'package:flutter_svg/flutter_svg.dart';
 
 /// A Flags.
 class Flags {
-  /// get mini flag
-  static Widget getMiniFlag(String country, double height, double width) {
-    String assetName = 'packages/flag/res/svg1/' + country + '.svg';
-    Widget svg = new SvgPicture.asset(
-      assetName,
-      semanticsLabel: country,
-      height: height,
+  /// get flag
+  static Widget getFlag(
+      {@required String country,
+      @required double height,
+      @required double width,
+      BoxFit fit = BoxFit.contain}) {
+    String countryName = country.toLowerCase();
+    String assetName = 'packages/flag/res/flag/' + countryName + '.svg';
+    return Container(
       width: width,
-    );
-    return svg;
-  }
-
-  /// get full flag
-  static Widget getFullFlag(String country, double height, double width) {
-    String assetName =
-        'packages/flag/res/svg2/' + country.toLowerCase() + '.svg';
-    Widget svg = new SvgPicture.asset(
-      assetName,
-      semanticsLabel: country.toLowerCase(),
       height: height,
-      width: width,
+      child: SvgPicture.asset(
+        assetName,
+        semanticsLabel: country,
+        fit: fit,
+      ),
     );
-    return svg;
   }
 }