+++ test/flag_test.dart
@@ -1,9 +1,9 @@
 import 'package:flutter_test/flutter_test.dart';
 
-import 'package:flag/flag.dart';
+// import 'package:flag/flag.dart';
 
 void main() {
   testWidgets('flag widget test', (WidgetTester tester) async {
-//    await tester.pumpWidget(Flags.getMiniFlag('AD'));
+    // await tester.pumpWidget(Flags.getFlag(country: 'AD'));
   });
 }