+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## [2.0.0] - Release Flags
+* list of all countries from https://www.un.org/en/member-states/index.html
+* flag from wiki.
+* update flutter_svg packages to ^0.17.4
+* remove `getFullFlag` and `getMiniFlag` function
+* add `getFlag` function
+* add fit option
+
 ## [1.0.3] - Release Flags
 
 * update flutter_svg packages to ^0.17.0